+++ CHANGELOG.md
@@ -16,6 +16,15 @@ Types of changes
 
 ## [Unreleased]
 
+## 1.1.1 - 2017-12-12
+
+### Added
+* Started using NanoHttpStatus to detect HTTP error status;
+* composer scripts.
+
+### Changed
+* Updated README.
+
 ## 1.1.0 - 2017-12-05
 
 ### Added
@@ -51,4 +60,4 @@ Types of changes
 * generation of docs failed - PhpDoc cannot build full class hierarchy;
 * more to come - short syntax for most common requests, cookies, response type based on content-type header, etc.
 
-[Unreleased]: https://github.com/GinoPane/php-nano-rest/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/GinoPane/php-nano-rest/compare/v1.1.1...HEAD